`dcat` dumps today's diary to stdout.
I find it useful to refresh my memory by reviewing today's pyd entries, but opening an editor every time is pretty disruptive. I wanted something more like `ls`, where the output is part of my normal terminal flow.

This adds a new script, `dcat`, that does exactly that. It takes no arguments or options (or rather, it ignores them) and either prints today's diary in pyd format on stdout, or prints an error on stderr if today has no dairy (and exits with a nonzero status).