Use ISO week numbers.
The year cutover last week revealed a bug in the "what is the current week
number" logic:

    * On Dec 31, the current week was 53-2012.
    * On Jan 1, the current week was 00-2013.
    * Both of these days fall on the same notional week.

I've switched from computing week number by string formatting to computing
week number using the ISO mechanism, which assigns the year based on the year
Thursday falls on (for reasons too dull to go into). This is a disruptive
change, as it renumbers all weeks in 2013 by one.